Sounds, usable indicator and saving scores
Some sounds are now usable in code, they just need to be called whenever they need.
Also, their volume must be modify when no more sounds will be added
Score data can now be saved in a file. Only the 5 best scores will be kept.
Whenever a bucket is in range of a player, a green "!" will spawn on top of it, showing the playing it can be used